Backend/bionic/requirements.txt
asgiref==3.8.1
distlib==0.3.7
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
django-rest-framework==0.1.0
djangorestframework==3.15.1
filelock==3.13.1
Markdown==3.6
pillow==10.2.0
platformdirs==3.11.0
sqlparse==0.4.4
tzdata==2024.1
virtualenv==20.24.6
